refactor(k8s): isolate PDS grant adapter
