+++ Cargo.toml
@@ -31,6 +31,7 @@ members = [
     "crates/hyprstream-workers-wasmtime",
     "crates/hyprstream-workers-python",
     "crates/hyprstream-k8s",
+    "crates/hyprstream-k8s-pds",
     "crates/hyprstream-ledger",
     "crates/hyprstream-resource",
 ]
