+++ crates/hyprstream-k8s-pds/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "hyprstream-k8s-pds"
+version = "0.1.0"
+edition.workspace = true
+license = "AGPL-3.0-only"
+repository.workspace = true
+description = "AGPL tenant-grant and PDS allocation adapter for hyprstream-k8s"
+
+[dependencies]
+anyhow = { workspace = true }
+ed25519-dalek = { workspace = true }
+hyprstream-k8s = { path = "../hyprstream-k8s", features = ["grant"] }
+hyprstream-pds = { path = "../hyprstream-pds" }
+hyprstream-rpc = { path = "../hyprstream-rpc" }
+rand = { workspace = true }
+
+[lints]
+workspace = true
